+++ glide-core/src/socket_listener.rs
@@ -348,6 +348,10 @@ fn get_command(request: &Command) -> Option<Cmd> {
         RequestType::ZScore => Some(cmd("ZSCORE")),
         RequestType::Type => Some(cmd("TYPE")),
         RequestType::HLen => Some(cmd("HLEN")),
+        RequestType::Echo => Some(cmd("ECHO")),
+        RequestType::ZPopMin => Some(cmd("ZPOPMIN")),
+        RequestType::Strlen => Some(cmd("STRLEN")),
+        RequestType::Lindex => Some(cmd("LINDEX")),
     }
 }