+++ java/integTest/src/test/java/glide/SharedCommandTests.java
@@ -7,6 +7,8 @@ import static glide.api.BaseClient.OK;
 import static glide.api.models.commands.SetOptions.ConditionalSet.ONLY_IF_DOES_NOT_EXIST;
 import static glide.api.models.commands.SetOptions.ConditionalSet.ONLY_IF_EXISTS;
 import static glide.api.models.commands.SetOptions.Expiry.Milliseconds;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -21,6 +23,7 @@ import glide.api.models.configuration.RedisClientConfiguration;
 import glide.api.models.configuration.RedisClusterClientConfiguration;
 import glide.api.models.exceptions.RequestException;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
@@ -251,6 +254,113 @@ public class SharedCommandTests {
         SetOptions options = SetOptions.builder().returnOldValue(true).build();
         String data = client.set("another", ANOTHER_VALUE, options).get();
         assertNull(data);
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void mset_mget_existing_non_existing_key(BaseClient client) {
+        String key1 = UUID.randomUUID().toString();
+        String key2 = UUID.randomUUID().toString();
+        String key3 = UUID.randomUUID().toString();
+        String nonExisting = UUID.randomUUID().toString();
+        String value = UUID.randomUUID().toString();
+        Map<String, String> keyValueMap = Map.of(key1, value, key2, value, key3, value);
+
+        assertEquals(OK, client.mset(keyValueMap).get());
+        assertArrayEquals(
+                new String[] {value, value, null, value},
+                client.mget(new String[] {key1, key2, nonExisting, key3}).get());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void incr_commands_existing_key(BaseClient client) {
+        String key = UUID.randomUUID().toString();
+
+        assertEquals(OK, client.set(key, "10").get());
+
+        assertEquals(11, client.incr(key).get());
+        assertEquals("11", client.get(key).get());
+
+        assertEquals(15, client.incrBy(key, 4).get());
+        assertEquals("15", client.get(key).get());
+
+        assertEquals(20.5, client.incrByFloat(key, 5.5).get());
+        assertEquals("20.5", client.get(key).get());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void incr_commands_non_existing_key(BaseClient client) {
+        String key1 = UUID.randomUUID().toString();
+        String key2 = UUID.randomUUID().toString();
+        String key3 = UUID.randomUUID().toString();
+
+        assertNull(client.get(key1).get());
+        assertEquals(1, client.incr(key1).get());
+        assertEquals("1", client.get(key1).get());
+
+        assertNull(client.get(key2).get());
+        assertEquals(3, client.incrBy(key2, 3).get());
+        assertEquals("3", client.get(key2).get());
+
+        assertNull(client.get(key3).get());
+        assertEquals(0.5, client.incrByFloat(key3, 0.5).get());
+        assertEquals("0.5", client.get(key3).get());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void test_incr_commands_type_error(BaseClient client) {
+        String key1 = UUID.randomUUID().toString();
+
+        assertEquals(OK, client.set(key1, "foo").get());
+
+        Exception incrException = assertThrows(ExecutionException.class, () -> client.incr(key1).get());
+        assertTrue(incrException.getCause() instanceof RequestException);
+
+        Exception incrByException =
+                assertThrows(ExecutionException.class, () -> client.incrBy(key1, 3).get());
+        assertTrue(incrByException.getCause() instanceof RequestException);
+
+        Exception incrByFloatException =
+                assertThrows(ExecutionException.class, () -> client.incrByFloat(key1, 3.5).get());
+        assertTrue(incrByFloatException.getCause() instanceof RequestException);
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void decr_and_decrBy_existing_key(BaseClient client) {
+        String key = UUID.randomUUID().toString();
+
+        assertEquals(OK, client.set(key, "10").get());
+
+        assertEquals(9, client.decr(key).get());
+        assertEquals("9", client.get(key).get());
+
+        assertEquals(5, client.decrBy(key, 4).get());
+        assertEquals("5", client.get(key).get());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @MethodSource("getClients")
+    public void decr_and_decrBy_non_existing_key(BaseClient client) {
+        String key1 = UUID.randomUUID().toString();
+        String key2 = UUID.randomUUID().toString();
+
+        assertNull(client.get(key1).get());
+        assertEquals(-1, client.decr(key1).get());
+        assertEquals("-1", client.get(key1).get());
+
+        assertNull(client.get(key2).get(10, SECONDS));
+        assertEquals(-3, client.decrBy(key2, 3).get());
+        assertEquals("-3", client.get(key2).get());
     }
 
     @SneakyThrows