+++ node/src/Commands.ts
@@ -823,3 +823,20 @@ export function createZcount(
 export function createType(key: string): redis_request.Command {
     return createCommand(RequestType.Type, [key]);
 }
+
+/**
+ * @internal
+ */
+export function createStrlen(key: string): redis_request.Command {
+    return createCommand(RequestType.Strlen, [key]);
+}
+
+/**
+ * @internal
+ */
+export function createLindex(
+    key: string,
+    index: number
+): redis_request.Command {
+    return createCommand(RequestType.Lindex, [key, index.toString()]);
+}