+++ node/tests/SharedTests.ts
@@ -1390,7 +1390,6 @@ export function runBaseTests<Context>(config: {
         config.timeout
     );
 
-
     it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
         `zscore test_%p`,
         async (protocol) => {
@@ -1506,6 +1505,50 @@ export function runBaseTests<Context>(config: {
                 expect(await client.del([key])).toEqual(1);
 
                 expect(await client.type(key)).toEqual("none");
+            }, protocol);
+        },
+        config.timeout
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `strlen test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key1 = uuidv4();
+                const key1Value = uuidv4();
+                const key1ValueLength = key1Value.length;
+                expect(await client.set(key1, key1Value)).toEqual("OK");
+                expect(await client.strlen(key1)).toEqual(key1ValueLength);
+
+                expect(await client.strlen("nonExistKey")).toEqual(0);
+
+                const listName = "myList";
+                const listKey1Value = uuidv4();
+                const listKey2Value = uuidv4();
+
+                expect(
+                    await client.lpush(listName, [listKey1Value, listKey2Value])
+                ).toEqual(2);
+                // An error is returned when key holds a non-string value
+                await expect(client.strlen(listName)).rejects.toThrow();
+            }, protocol);
+        },
+        config.timeout
+    );
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `lindex test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const listName = uuidv4();
+                const listKey1Value = uuidv4();
+                const listKey2Value = uuidv4();
+                expect(
+                    await client.lpush(listName, [listKey1Value, listKey2Value])
+                ).toEqual(2);
+                expect(await client.lindex(listName, 0)).toEqual(listKey2Value);
+                expect(await client.lindex(listName, 1)).toEqual(listKey1Value);
+                expect(await client.lindex("notExsitingList", 1)).toEqual(null);
+                expect(await client.lindex(listName, 3)).toEqual(null);
             }, protocol);
         },
         config.timeout