+++ CHANGELOG.md
@@ -12,10 +12,14 @@
 * Python, Node: Added TYPE command ([#945](https://github.com/aws/glide-for-redis/pull/945), [#980](https://github.com/aws/glide-for-redis/pull/980))
 * Python: Added HLEN command ([#944](https://github.com/aws/glide-for-redis/pull/944))
 * Node: Added ZCOUNT command ([#909](https://github.com/aws/glide-for-redis/pull/909))
+* Python: Added ECHO command ([#953](https://github.com/aws/glide-for-redis/pull/953))
+* Python: Added ZPOPMIN command ([#975](https://github.com/aws/glide-for-redis/pull/975))
+* Node: Added STRLEN command ([#993](https://github.com/aws/glide-for-redis/pull/993))
+* Node: Added LINDEX command ([#999](https://github.com/aws/glide-for-redis/pull/999))
 
 #### Features
 * Python, Node: Added support in Lua Scripts ([#775](https://github.com/aws/glide-for-redis/pull/775), [#860](https://github.com/aws/glide-for-redis/pull/860))
-* Node: Allow chaining function calls on transaction. ([#902](https://github.com/aws/glide-for-redis/pull/902)) 
+* Python, Node: Allow chaining function calls on transaction. ([#902](https://github.com/aws/glide-for-redis/pull/902)), ([#987](https://github.com/aws/glide-for-redis/pull/987))
 
 #### Fixes
 * Core: Fixed `Connection Refused` error not to close the client ([#872](https://github.com/aws/glide-for-redis/pull/872))